FIX CDR terminati

+++ src/main/java/it/cnr/rsi/domain/Cdr.java
@@ -36,10 +36,10 @@ public class Cdr implements Serializable {
 	private Date duva;
 
 	@Column(name="ESERCIZIO_FINE")
-	private BigDecimal esercizioFine;
+	private Integer esercizioFine;
 
 	@Column(name="ESERCIZIO_INIZIO")
-	private BigDecimal esercizioInizio;
+	private Integer esercizioInizio;
 
 	private String indirizzo;
 
@@ -117,19 +117,19 @@ public class Cdr implements Serializable {
 		this.duva = duva;
 	}
 
-	public BigDecimal getEsercizioFine() {
+	public Integer getEsercizioFine() {
 		return this.esercizioFine;
 	}
 
-	public void setEsercizioFine(BigDecimal esercizioFine) {
+	public void setEsercizioFine(Integer esercizioFine) {
 		this.esercizioFine = esercizioFine;
 	}
 
-	public BigDecimal getEsercizioInizio() {
+	public Integer getEsercizioInizio() {
 		return this.esercizioInizio;
 	}
 
-	public void setEsercizioInizio(BigDecimal esercizioInizio) {
+	public void setEsercizioInizio(Integer esercizioInizio) {
 		this.esercizioInizio = esercizioInizio;
 	}
